[EI-999] [web] refactor: move file folder to function-call-handler folder
